+++ CSharp_Assignments_Phase2/CSharp_Assignments_Phase2.sln
@@ -5,7 +5,7 @@ VisualStudioVersion = 17.7.34003.232
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Multithreading", "Multithreading\Multithreading.csproj", "{C3E31D5D-C8FD-4749-BFDA-8148C71BFD21}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "MoneyManagerConsoleApplication", "MoneyManagerConsoleApplication\MoneyManagerConsoleApplication.csproj", "{7D1EE50F-FC7E-4E1E-BD65-AF77878A177C}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ExpenseTrackerConsoleApplication", "ExpenseTrackerConsoleApplication\ExpenseTrackerConsoleApplication.csproj", "{E63205BC-0126-42A2-B576-43D41546D275}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -17,10 +17,10 @@ Global
 		{C3E31D5D-C8FD-4749-BFDA-8148C71BFD21}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{C3E31D5D-C8FD-4749-BFDA-8148C71BFD21}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{C3E31D5D-C8FD-4749-BFDA-8148C71BFD21}.Release|Any CPU.Build.0 = Release|Any CPU
-		{7D1EE50F-FC7E-4E1E-BD65-AF77878A177C}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{7D1EE50F-FC7E-4E1E-BD65-AF77878A177C}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{7D1EE50F-FC7E-4E1E-BD65-AF77878A177C}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{7D1EE50F-FC7E-4E1E-BD65-AF77878A177C}.Release|Any CPU.Build.0 = Release|Any CPU
+		{E63205BC-0126-42A2-B576-43D41546D275}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{E63205BC-0126-42A2-B576-43D41546D275}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{E63205BC-0126-42A2-B576-43D41546D275}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{E63205BC-0126-42A2-B576-43D41546D275}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
