+++ CSharp_Assignments_Phase2/ExpenseTrackerConsoleApplication/ActiveUsers.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace MoneyManagerConsoleApplication
+namespace ExpenseTrackerConsoleApplication
 {
     /// <summary>
     /// Users that are active currently.
@@ -14,8 +14,6 @@ namespace MoneyManagerConsoleApplication
         /// <summary>
         /// List of users that are active.
         /// </summary>
-#pragma warning disable SA1401 // Fields should be private
         public static List<User> ActiveUsersList = new List<User>();
-#pragma warning restore SA1401 // Fields should be private
     }
 }
