+++ CSharp_Assignments_Phase2/ExpenseTrackerConsoleApplication/LogOut.cs
@@ -4,14 +4,24 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace MoneyManagerConsoleApplication
+namespace ExpenseTrackerConsoleApplication
 {
-    internal class SignOut
+    /// <summary>
+    /// Represents User LogOut.
+    /// </summary>
+    internal class LogOut
     {
         Parser parser = new Parser();
-        public async Task SignOutFromUserList(User user)
+
+        /// <summary>
+        /// User LogOut.
+        /// </summary>
+        /// <param name="userName">UserName</param>
+        /// <param name="password">Password</param>
+        /// <returns>Task</returns>
+        public async Task LogOutUser(User user)
         {
-            if (ActiveUsers.ActiveUsersList.Contains(user))
+            if(ActiveUsers.ActiveUsersList.Contains(user))
             {
                 int i = 0;
                 await Task.Run(() =>
@@ -23,7 +33,7 @@ namespace MoneyManagerConsoleApplication
                         i++;
                     }
                     ActiveUsers.ActiveUsersList.Remove(user);
-                    SignIn.users.Remove(user.UserName);
+                    parser.DisplayMessages($"Successfully Logged out ! \n ThankYou {user.UserName}", 5);
 
                     Thread.Sleep(100);
                 });
@@ -33,7 +43,7 @@ namespace MoneyManagerConsoleApplication
             {
                 parser.DisplayMessages("No User found !");
             }
-
+            
             return;
         }
     }
