+++ CSharp_Assignments_Phase2/ExpenseTrackerConsoleApplication/ModesOfCash.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace MoneyManagerConsoleApplication
+namespace ExpenseTrackerConsoleApplication
 {
     /// <summary>
     /// Modes of cash.
