+++ CSharp_Assignments_Phase2/ExpenseTrackerConsoleApplication/Program.cs
@@ -1,7 +1,14 @@
-﻿namespace MoneyManagerConsoleApplication
+﻿namespace ExpenseTrackerConsoleApplication
 {
+    /// <summary>
+    /// Driver class.
+    /// </summary>
     internal class Program
     {
+        /// <summary>
+        /// Driver method.
+        /// </summary>
+        /// <returns>Task</returns>
         public static async Task Main()
         {
             try
@@ -13,7 +20,7 @@
             {
                 Console.WriteLine(ex.ToString());
             }
-            
+
         }
     }
 }
