--- CSharp_Assignments_Phase2/MoneyManagerConsoleApplication/UpdateAttributes.cs
@@ -1,18 +1,0 @@
-﻿namespace MoneyManagerConsoleApplication
-{
-    internal partial class User
-    {
-        /// <summary>
-        /// Enum
-        /// </summary>
-        public enum UpdateAttributes
-        {
-            Date = 1,
-            Amount = 2,
-            AmountMode = 3,
-            Category = 4,
-            Account = 5,
-            Notes = 6,
-        }
-    }
-}
